Refactor query apply helpers into focused files
