+++ internal/query/applier.go
@@ -1,109 +1,13 @@
 package query
 
 import (
-	"fmt"
-	"reflect"
-	"strings"
-
 	"github.com/nlstn/go-odata/internal/metadata"
 	"gorm.io/gorm"
-	"gorm.io/gorm/clause"
 )
 
 // ShouldUseMapResults returns true if the query options require map results instead of entity results
 func ShouldUseMapResults(options *QueryOptions) bool {
 	return options != nil && (len(options.Apply) > 0 || options.Compute != nil)
-}
-
-// applySelectToExpandedEntity applies select to an expanded navigation property (single entity or collection)
-// This is a simplified version that doesn't require full metadata - it works with reflection
-func applySelectToExpandedEntity(expandedValue interface{}, selectedProps []string) interface{} {
-	if len(selectedProps) == 0 || expandedValue == nil {
-		return expandedValue
-	}
-
-	// Build a map of selected properties for quick lookup
-	selectedPropMap := make(map[string]bool)
-	for _, propName := range selectedProps {
-		selectedPropMap[strings.TrimSpace(propName)] = true
-	}
-
-	val := reflect.ValueOf(expandedValue)
-
-	// Handle pointer
-	if val.Kind() == reflect.Ptr {
-		if val.IsNil() {
-			return expandedValue
-		}
-		val = val.Elem()
-	}
-
-	// Handle slice/array (collection navigation)
-	if val.Kind() == reflect.Slice || val.Kind() == reflect.Array {
-		resultSlice := make([]map[string]interface{}, val.Len())
-		for i := 0; i < val.Len(); i++ {
-			itemVal := val.Index(i)
-			resultSlice[i] = filterEntityFields(itemVal, selectedPropMap)
-		}
-		return resultSlice
-	}
-
-	// Handle single entity
-	if val.Kind() == reflect.Struct {
-		return filterEntityFields(val, selectedPropMap)
-	}
-
-	return expandedValue
-}
-
-// filterEntityFields filters struct fields based on selected properties map
-func filterEntityFields(entityVal reflect.Value, selectedPropMap map[string]bool) map[string]interface{} {
-	filtered := make(map[string]interface{})
-	entityType := entityVal.Type()
-
-	// Always include ID/key fields
-	idFields := []string{"ID", "Id", "id"}
-
-	for i := 0; i < entityVal.NumField(); i++ {
-		field := entityType.Field(i)
-		fieldVal := entityVal.Field(i)
-
-		if !fieldVal.CanInterface() {
-			continue
-		}
-
-		// Get JSON name from tag if available
-		jsonName := field.Name
-		if jsonTag := field.Tag.Get("json"); jsonTag != "" {
-			parts := strings.Split(jsonTag, ",")
-			if parts[0] != "" && parts[0] != "-" {
-				jsonName = parts[0]
-			}
-		}
-
-		// Check if this field is selected or is an ID field
-		isSelected := selectedPropMap[field.Name] || selectedPropMap[jsonName]
-		isKeyField := false
-		for _, keyName := range idFields {
-			if field.Name == keyName {
-				isKeyField = true
-				break
-			}
-		}
-
-		// Also check for odata:"key" tag
-		if odataTag := field.Tag.Get("odata"); odataTag != "" {
-			if strings.Contains(odataTag, "key") {
-				isKeyField = true
-			}
-		}
-
-		if isSelected || isKeyField {
-			filtered[jsonName] = fieldVal.Interface()
-		}
-	}
-
-	return filtered
 }
 
 // ApplyQueryOptions applies parsed query options to a GORM database query
@@ -187,1659 +91,4 @@ func ApplyExpandOnly(db *gorm.DB, expand []ExpandOption, entityMetadata *metadat
 		return db
 	}
 	return applyExpand(db, expand, entityMetadata)
-}
-
-// applySelect applies select clause to fetch only specified columns at database level
-func applySelect(db *gorm.DB, selectedProperties []string, entityMetadata *metadata.EntityMetadata) *gorm.DB {
-	if len(selectedProperties) == 0 {
-		return db
-	}
-
-	// Build list of columns to select using struct field names
-	// GORM will convert them to the correct column names
-	columns := make([]string, 0, len(selectedProperties))
-	selectedPropMap := make(map[string]bool)
-
-	// Add requested properties
-	for _, propName := range selectedProperties {
-		propName = strings.TrimSpace(propName)
-		// Find the property in metadata
-		for _, prop := range entityMetadata.Properties {
-			// Skip navigation properties, complex types, and stream properties - they are handled separately
-			if (prop.JsonName == propName || prop.Name == propName) && !prop.IsNavigationProp && !prop.IsComplexType && !prop.IsStream {
-				// Use struct field name - GORM will handle column name conversion
-				columns = append(columns, prop.Name)
-				selectedPropMap[prop.Name] = true
-				break
-			}
-		}
-	}
-
-	// Always include key properties for OData responses (if not already selected)
-	for _, keyProp := range entityMetadata.KeyProperties {
-		if !selectedPropMap[keyProp.Name] {
-			columns = append(columns, keyProp.Name)
-		}
-	}
-
-	// Apply select to GORM query
-	if len(columns) > 0 {
-		db = db.Select(columns)
-	}
-
-	return db
-}
-
-// applyFilter applies filter expressions to the GORM query
-func applyFilter(db *gorm.DB, filter *FilterExpression, entityMetadata *metadata.EntityMetadata) *gorm.DB {
-	if filter == nil {
-		return db
-	}
-
-	// Handle logical operators (and, or)
-	if filter.Logical != "" {
-		// Build the complete condition for logical operators
-		query, args := buildFilterCondition(filter, entityMetadata)
-		return db.Where(query, args...)
-	}
-
-	// Handle simple comparison operators
-	query, args := buildFilterCondition(filter, entityMetadata)
-	return db.Where(query, args...)
-}
-
-// applyHavingFilter applies a HAVING clause filter for post-aggregation filtering
-// This is used when a filter transformation comes after a groupby/aggregate in $apply
-func applyHavingFilter(db *gorm.DB, filter *FilterExpression, entityMetadata *metadata.EntityMetadata) *gorm.DB {
-	if filter == nil {
-		return db
-	}
-
-	// Build the condition using the same logic as WHERE, but apply as HAVING
-	query, args := buildFilterCondition(filter, entityMetadata)
-	return db.Having(query, args...)
-}
-
-// buildFilterCondition builds a WHERE condition string and arguments for a filter expression
-func buildFilterCondition(filter *FilterExpression, entityMetadata *metadata.EntityMetadata) (string, []interface{}) {
-	if filter == nil {
-		return "", nil
-	}
-
-	// Handle logical operators recursively
-	if filter.Logical != "" {
-		return buildLogicalCondition(filter, entityMetadata)
-	}
-
-	// Build comparison condition
-	query, args := buildComparisonCondition(filter, entityMetadata)
-
-	// Apply NOT if needed
-	if filter.IsNot {
-		return fmt.Sprintf("NOT (%s)", query), args
-	}
-
-	return query, args
-}
-
-// buildLogicalCondition builds a logical condition (AND/OR)
-func buildLogicalCondition(filter *FilterExpression, entityMetadata *metadata.EntityMetadata) (string, []interface{}) {
-	leftQuery, leftArgs := buildFilterCondition(filter.Left, entityMetadata)
-	rightQuery, rightArgs := buildFilterCondition(filter.Right, entityMetadata)
-
-	var query string
-	switch filter.Logical {
-	case LogicalAnd:
-		query = fmt.Sprintf("(%s) AND (%s)", leftQuery, rightQuery)
-	case LogicalOr:
-		query = fmt.Sprintf("(%s) OR (%s)", leftQuery, rightQuery)
-	default:
-		return "", nil
-	}
-
-	args := append(leftArgs, rightArgs...)
-	if filter.IsNot {
-		return fmt.Sprintf("NOT (%s)", query), args
-	}
-	return query, args
-}
-
-// buildComparisonCondition builds a comparison condition
-func buildComparisonCondition(filter *FilterExpression, entityMetadata *metadata.EntityMetadata) (string, []interface{}) {
-	// Check if this is a lambda expression (any/all)
-	if filter.Operator == OpAny || filter.Operator == OpAll {
-		return buildLambdaCondition(filter, entityMetadata)
-	}
-
-	// Check if this is a function comparison (function on left side of comparison)
-	if filter.Left != nil && filter.Left.Operator != "" {
-		return buildFunctionComparison(filter, entityMetadata)
-	}
-
-	// Determine if this is a regular entity property or a computed property
-	// If the property doesn't exist in entity metadata, treat it as a computed alias
-	var columnName string
-	if propertyExists(filter.Property, entityMetadata) {
-		columnName = GetColumnName(filter.Property, entityMetadata)
-	} else {
-		// Computed property - sanitize the alias to prevent SQL injection
-		rawName := sanitizeIdentifier(filter.Property)
-		if rawName == "" {
-			// Invalid identifier - return empty condition
-			return "", nil
-		}
-		// Handle special OData identifiers
-		if rawName == "$it" {
-			// $it refers to the current instance, don't quote
-			columnName = rawName
-		} else if rawName == "$count" {
-			// $count is a computed column alias, must be quoted to avoid being treated as SQL placeholder
-			columnName = "`" + rawName + "`"
-		} else if len(rawName) > 0 && rawName[0] == '$' {
-			// Quote other identifiers that start with $ to avoid conflicts with SQL placeholders
-			columnName = "`" + rawName + "`"
-		} else {
-			columnName = rawName
-		}
-	}
-
-	// Check if the value is a function call that needs to be converted to SQL
-	if funcCall, ok := filter.Value.(*FunctionCallExpr); ok {
-		// Convert the function call to a filter expression
-		rightExpr, err := convertFunctionCallExpr(funcCall, entityMetadata)
-		if err == nil && rightExpr != nil {
-			// Build SQL for the function
-			rightColumnName := ""
-			if rightExpr.Property != "" {
-				rightColumnName = GetColumnName(rightExpr.Property, entityMetadata)
-			}
-			rightSQL, rightArgs := buildFunctionSQL(rightExpr.Operator, rightColumnName, rightExpr.Value)
-			if rightSQL != "" {
-				// Build comparison with the function result
-				var compSQL string
-				switch filter.Operator {
-				case OpEqual:
-					compSQL = fmt.Sprintf("%s = %s", columnName, rightSQL)
-				case OpNotEqual:
-					compSQL = fmt.Sprintf("%s != %s", columnName, rightSQL)
-				case OpGreaterThan:
-					compSQL = fmt.Sprintf("%s > %s", columnName, rightSQL)
-				case OpGreaterThanOrEqual:
-					compSQL = fmt.Sprintf("%s >= %s", columnName, rightSQL)
-				case OpLessThan:
-					compSQL = fmt.Sprintf("%s < %s", columnName, rightSQL)
-				case OpLessThanOrEqual:
-					compSQL = fmt.Sprintf("%s <= %s", columnName, rightSQL)
-				default:
-					return "", nil
-				}
-				return compSQL, rightArgs
-			}
-		}
-	}
-
-	switch filter.Operator {
-	case OpEqual:
-		// Handle null comparisons with IS NULL
-		if filter.Value == nil {
-			return fmt.Sprintf("%s IS NULL", columnName), []interface{}{}
-		}
-		return fmt.Sprintf("%s = ?", columnName), []interface{}{filter.Value}
-	case OpNotEqual:
-		// Handle null comparisons with IS NOT NULL
-		if filter.Value == nil {
-			return fmt.Sprintf("%s IS NOT NULL", columnName), []interface{}{}
-		}
-		return fmt.Sprintf("%s != ?", columnName), []interface{}{filter.Value}
-	case OpGreaterThan:
-		return fmt.Sprintf("%s > ?", columnName), []interface{}{filter.Value}
-	case OpGreaterThanOrEqual:
-		return fmt.Sprintf("%s >= ?", columnName), []interface{}{filter.Value}
-	case OpLessThan:
-		return fmt.Sprintf("%s < ?", columnName), []interface{}{filter.Value}
-	case OpLessThanOrEqual:
-		return fmt.Sprintf("%s <= ?", columnName), []interface{}{filter.Value}
-	case OpIn:
-		// Handle IN operator with a collection of values
-		values, ok := filter.Value.([]interface{})
-		if !ok {
-			return "", nil
-		}
-		if len(values) == 0 {
-			// Empty IN clause - return false condition
-			return "1 = 0", []interface{}{}
-		}
-		placeholders := make([]string, len(values))
-		for i := range values {
-			placeholders[i] = "?"
-		}
-		return fmt.Sprintf("%s IN (%s)", columnName, strings.Join(placeholders, ", ")), values
-	case OpContains:
-		return fmt.Sprintf("%s LIKE ?", columnName), []interface{}{"%" + fmt.Sprint(filter.Value) + "%"}
-	case OpStartsWith:
-		return fmt.Sprintf("%s LIKE ?", columnName), []interface{}{fmt.Sprint(filter.Value) + "%"}
-	case OpEndsWith:
-		return fmt.Sprintf("%s LIKE ?", columnName), []interface{}{"%" + fmt.Sprint(filter.Value)}
-	case OpHas:
-		// Bitwise AND for enum flags: (column & value) = value
-		return fmt.Sprintf("(%s & ?) = ?", columnName), []interface{}{filter.Value, filter.Value}
-	case OpIsOf:
-		// Handle isof as a standalone boolean expression
-		// This generates SQL that can be used directly in a WHERE clause
-		funcSQL, funcArgs := buildFunctionSQL(OpIsOf, columnName, filter.Value)
-		if funcSQL == "" {
-			return "", nil
-		}
-		// isof returns 1 (true) or 0 (false), so we compare it with true (1)
-		return fmt.Sprintf("%s = ?", funcSQL), append(funcArgs, true)
-	case OpGeoIntersects:
-		// Handle geo.intersects as a standalone boolean expression
-		funcSQL, funcArgs := buildFunctionSQL(OpGeoIntersects, columnName, filter.Value)
-		if funcSQL == "" {
-			return "", nil
-		}
-		// ST_Intersects returns true/false (1/0), so we can use it directly in WHERE
-		return funcSQL, funcArgs
-	case OpCast:
-		// Cast alone doesn't make sense as a boolean expression
-		// It should always be part of a comparison
-		return "", nil
-	default:
-		return "", nil
-	}
-}
-
-// buildLambdaCondition builds SQL for lambda operators (any/all) using EXISTS subquery
-func buildLambdaCondition(filter *FilterExpression, entityMetadata *metadata.EntityMetadata) (string, []interface{}) {
-	// Extract navigation property info
-	navProp := findNavigationProperty(filter.Property, entityMetadata)
-	if navProp == nil {
-		// Not a navigation property, can't apply lambda
-		return "", nil
-	}
-
-	// Get the related entity's table name using GORM conventions
-	// NavigationTarget is the entity type name (e.g., "ProductDescription")
-	// GORM table name would be snake_case + pluralized (e.g., "product_descriptions")
-	relatedEntityName := navProp.NavigationTarget
-	if relatedEntityName == "" {
-		// Fallback to using the property name
-		relatedEntityName = navProp.JsonName
-	}
-	relatedTableName := toSnakeCase(pluralize(relatedEntityName))
-
-	// Get the parent entity's table name
-	parentTableName := toSnakeCase(pluralize(entityMetadata.EntityName))
-
-	// Get the foreign key column name (usually parent_id)
-	// For ProductDescription -> Product, foreign key is product_id
-	foreignKeyColumn := toSnakeCase(entityMetadata.EntityName) + "_id"
-
-	// Get the parent's primary key column name (usually "id")
-	parentPrimaryKey := "id"
-	if len(entityMetadata.KeyProperties) > 0 {
-		parentPrimaryKey = toSnakeCase(entityMetadata.KeyProperties[0].Name)
-	}
-
-	// Handle parameterless any/all
-	if filter.Value == nil || filter.Left == nil {
-		// Parameterless: just check if collection has any elements
-		if filter.Operator == OpAny {
-			// any() - at least one related record exists
-			return fmt.Sprintf("EXISTS (SELECT 1 FROM %s WHERE %s.%s = %s.%s)",
-				relatedTableName, relatedTableName, foreignKeyColumn, parentTableName, parentPrimaryKey), []interface{}{}
-		}
-		// all() without predicate doesn't make semantic sense in OData but we handle it
-		// It would mean "all elements satisfy true" which is always true if collection exists
-		return fmt.Sprintf("NOT EXISTS (SELECT 1 FROM %s WHERE %s.%s = %s.%s) OR EXISTS (SELECT 1 FROM %s WHERE %s.%s = %s.%s)",
-			relatedTableName, relatedTableName, foreignKeyColumn, parentTableName, parentPrimaryKey,
-			relatedTableName, relatedTableName, foreignKeyColumn, parentTableName, parentPrimaryKey), []interface{}{}
-	}
-
-	// Extract predicate from filter.Left
-	predicate := filter.Left
-	if predicate == nil {
-		return "", nil
-	}
-
-	// Build the condition for the subquery based on the predicate
-	// The predicate refers to properties of the related entity
-	predicateSQL, predicateArgs := buildFilterConditionForLambda(predicate, navProp)
-	if predicateSQL == "" {
-		return "", nil
-	}
-
-	// Build the complete EXISTS subquery
-	var sql string
-	if filter.Operator == OpAny {
-		// any(predicate) - at least one related record satisfies the predicate
-		sql = fmt.Sprintf("EXISTS (SELECT 1 FROM %s WHERE %s.%s = %s.%s AND %s)",
-			relatedTableName, relatedTableName, foreignKeyColumn, parentTableName, parentPrimaryKey, predicateSQL)
-	} else {
-		// all(predicate) - all related records satisfy the predicate
-		// This is equivalent to: NOT EXISTS (related record that does NOT satisfy predicate)
-		sql = fmt.Sprintf("NOT EXISTS (SELECT 1 FROM %s WHERE %s.%s = %s.%s AND NOT (%s))",
-			relatedTableName, relatedTableName, foreignKeyColumn, parentTableName, parentPrimaryKey, predicateSQL)
-	}
-
-	return sql, predicateArgs
-}
-
-// buildFilterConditionForLambda builds a filter condition for lambda subquery predicates
-// This is similar to buildFilterCondition but works with the related entity's metadata
-func buildFilterConditionForLambda(filter *FilterExpression, navProp *metadata.PropertyMetadata) (string, []interface{}) {
-	if filter == nil {
-		return "", nil
-	}
-
-	// Handle logical operators recursively
-	if filter.Logical != "" {
-		return buildLogicalConditionForLambda(filter, navProp)
-	}
-
-	// Handle function comparisons
-	if filter.Left != nil && filter.Left.Operator != "" {
-		return buildFunctionComparisonForLambda(filter, navProp)
-	}
-
-	// Simple comparison - convert property to column name for the related entity
-	columnName := toSnakeCase(filter.Property)
-
-	switch filter.Operator {
-	case OpEqual:
-		if filter.Value == nil {
-			return fmt.Sprintf("%s IS NULL", columnName), []interface{}{}
-		}
-		return fmt.Sprintf("%s = ?", columnName), []interface{}{filter.Value}
-	case OpNotEqual:
-		if filter.Value == nil {
-			return fmt.Sprintf("%s IS NOT NULL", columnName), []interface{}{}
-		}
-		return fmt.Sprintf("%s != ?", columnName), []interface{}{filter.Value}
-	case OpGreaterThan:
-		return fmt.Sprintf("%s > ?", columnName), []interface{}{filter.Value}
-	case OpGreaterThanOrEqual:
-		return fmt.Sprintf("%s >= ?", columnName), []interface{}{filter.Value}
-	case OpLessThan:
-		return fmt.Sprintf("%s < ?", columnName), []interface{}{filter.Value}
-	case OpLessThanOrEqual:
-		return fmt.Sprintf("%s <= ?", columnName), []interface{}{filter.Value}
-	case OpContains:
-		return fmt.Sprintf("%s LIKE ?", columnName), []interface{}{"%" + fmt.Sprint(filter.Value) + "%"}
-	case OpStartsWith:
-		return fmt.Sprintf("%s LIKE ?", columnName), []interface{}{fmt.Sprint(filter.Value) + "%"}
-	case OpEndsWith:
-		return fmt.Sprintf("%s LIKE ?", columnName), []interface{}{"%" + fmt.Sprint(filter.Value)}
-	default:
-		return "", nil
-	}
-}
-
-// buildLogicalConditionForLambda builds a logical condition for lambda predicates
-func buildLogicalConditionForLambda(filter *FilterExpression, navProp *metadata.PropertyMetadata) (string, []interface{}) {
-	leftQuery, leftArgs := buildFilterConditionForLambda(filter.Left, navProp)
-	rightQuery, rightArgs := buildFilterConditionForLambda(filter.Right, navProp)
-
-	var query string
-	switch filter.Logical {
-	case LogicalAnd:
-		query = fmt.Sprintf("(%s) AND (%s)", leftQuery, rightQuery)
-	case LogicalOr:
-		query = fmt.Sprintf("(%s) OR (%s)", leftQuery, rightQuery)
-	default:
-		return "", nil
-	}
-
-	args := append(leftArgs, rightArgs...)
-	return query, args
-}
-
-// buildFunctionComparisonForLambda builds a function comparison for lambda predicates
-func buildFunctionComparisonForLambda(filter *FilterExpression, _ *metadata.PropertyMetadata) (string, []interface{}) {
-	funcExpr := filter.Left
-	columnName := toSnakeCase(funcExpr.Property)
-
-	funcSQL, funcArgs := buildFunctionSQL(funcExpr.Operator, columnName, funcExpr.Value)
-	if funcSQL == "" {
-		return "", nil
-	}
-
-	// Build comparison
-	compSQL := buildComparisonSQL(filter.Operator, funcSQL)
-	if compSQL == "" {
-		return "", nil
-	}
-
-	allArgs := append(funcArgs, filter.Value)
-	return compSQL, allArgs
-}
-
-// buildFunctionComparison builds a comparison with a function on the left side
-func buildFunctionComparison(filter *FilterExpression, entityMetadata *metadata.EntityMetadata) (string, []interface{}) {
-	funcExpr := filter.Left
-	columnName := GetColumnName(funcExpr.Property, entityMetadata)
-
-	funcSQL, funcArgs := buildFunctionSQL(funcExpr.Operator, columnName, funcExpr.Value)
-	if funcSQL == "" {
-		return "", nil
-	}
-
-	// Check if right side is also a function call
-	if rightFunc, ok := filter.Value.(*FunctionCallExpr); ok {
-		// Convert the right-side function call to a FilterExpression
-		rightFuncExpr, err := convertFunctionCallExpr(rightFunc, entityMetadata)
-		if err != nil {
-			// If conversion fails, fall back to treating it as a literal
-			compSQL := buildComparisonSQL(filter.Operator, funcSQL)
-			if compSQL == "" {
-				return "", nil
-			}
-			allArgs := append(funcArgs, filter.Value)
-			return compSQL, allArgs
-		}
-
-		// Build SQL for the right-side function
-		rightColumnName := GetColumnName(rightFuncExpr.Property, entityMetadata)
-		rightFuncSQL, rightFuncArgs := buildFunctionSQL(rightFuncExpr.Operator, rightColumnName, rightFuncExpr.Value)
-		if rightFuncSQL == "" {
-			return "", nil
-		}
-
-		// Build comparison between two functions
-		var compSQL string
-		switch filter.Operator {
-		case OpEqual:
-			compSQL = fmt.Sprintf("%s = %s", funcSQL, rightFuncSQL)
-		case OpNotEqual:
-			compSQL = fmt.Sprintf("%s != %s", funcSQL, rightFuncSQL)
-		case OpGreaterThan:
-			compSQL = fmt.Sprintf("%s > %s", funcSQL, rightFuncSQL)
-		case OpGreaterThanOrEqual:
-			compSQL = fmt.Sprintf("%s >= %s", funcSQL, rightFuncSQL)
-		case OpLessThan:
-			compSQL = fmt.Sprintf("%s < %s", funcSQL, rightFuncSQL)
-		case OpLessThanOrEqual:
-			compSQL = fmt.Sprintf("%s <= %s", funcSQL, rightFuncSQL)
-		default:
-			return "", nil
-		}
-
-		allArgs := append(funcArgs, rightFuncArgs...)
-		return compSQL, allArgs
-	}
-
-	// Right side is a literal or property
-	compSQL := buildComparisonSQL(filter.Operator, funcSQL)
-	if compSQL == "" {
-		return "", nil
-	}
-
-	allArgs := append(funcArgs, filter.Value)
-	return compSQL, allArgs
-}
-
-// buildFunctionSQL generates SQL for a string function
-func buildFunctionSQL(op FilterOperator, columnName string, value interface{}) (string, []interface{}) {
-	switch op {
-	case OpToLower:
-		return fmt.Sprintf("LOWER(%s)", columnName), nil
-	case OpToUpper:
-		return fmt.Sprintf("UPPER(%s)", columnName), nil
-	case OpTrim:
-		return fmt.Sprintf("TRIM(%s)", columnName), nil
-	case OpLength:
-		return fmt.Sprintf("LENGTH(%s)", columnName), nil
-	case OpIndexOf:
-		// OData indexof is 0-indexed and returns -1 when not found
-		// SQLite INSTR is 1-indexed and returns 0 when not found
-		// Converting: INSTR - 1 gives us the correct OData indexof behavior
-		return fmt.Sprintf("INSTR(%s, ?) - 1", columnName), []interface{}{value}
-	case OpConcat:
-		// Check if value is a slice (both arguments stored)
-		if args, ok := value.([]interface{}); ok && len(args) == 2 {
-			// Both arguments provided (first is literal or function)
-			firstArg := args[0]
-			secondArg := args[1]
-
-			// Build SQL for both arguments
-			var leftSQL string
-			var leftArgs []interface{}
-
-			// Handle first argument
-			if funcCall, ok := firstArg.(*FunctionCallExpr); ok {
-				// First argument is a function call
-				leftExpr, err := convertFunctionCallExpr(funcCall, nil)
-				if err == nil {
-					leftSQL, leftArgs = buildFunctionSQL(leftExpr.Operator, leftExpr.Property, leftExpr.Value)
-				}
-				if leftSQL == "" {
-					leftSQL = "?"
-					leftArgs = []interface{}{firstArg}
-				}
-			} else {
-				// First argument is a literal
-				leftSQL = "?"
-				leftArgs = []interface{}{firstArg}
-			}
-
-			// Handle second argument
-			var rightSQL string
-			var rightArgs []interface{}
-
-			if funcCall, ok := secondArg.(*FunctionCallExpr); ok {
-				// Second argument is a function call
-				rightExpr, err := convertFunctionCallExpr(funcCall, nil)
-				if err == nil {
-					rightSQL, rightArgs = buildFunctionSQL(rightExpr.Operator, rightExpr.Property, rightExpr.Value)
-				}
-				if rightSQL == "" {
-					rightSQL = "?"
-					rightArgs = []interface{}{secondArg}
-				}
-			} else if strVal, ok := secondArg.(string); ok {
-				// Check if it's a property name or literal
-				if len(strVal) > 0 && (strVal[0] >= 'A' && strVal[0] <= 'Z' || strings.Contains(strVal, "_")) {
-					// Treat as column name
-					rightSQL = toSnakeCase(strVal)
-					rightArgs = nil
-				} else {
-					// Treat as literal
-					rightSQL = "?"
-					rightArgs = []interface{}{strVal}
-				}
-			} else {
-				// Other literal type
-				rightSQL = "?"
-				rightArgs = []interface{}{secondArg}
-			}
-
-			// Combine arguments
-			allArgs := append(leftArgs, rightArgs...)
-			return fmt.Sprintf("CONCAT(%s, %s)", leftSQL, rightSQL), allArgs
-		}
-
-		// Standard case: first argument is a property (in columnName)
-		// Handle different types of second argument
-		if funcCall, ok := value.(*FunctionCallExpr); ok {
-			// Second argument is a function call - need to convert it
-			rightExpr, err := convertFunctionCallExpr(funcCall, nil)
-			if err != nil {
-				return fmt.Sprintf("CONCAT(%s, ?)", columnName), []interface{}{value}
-			}
-			// Build SQL for the nested function
-			rightColumnName := rightExpr.Property
-			rightSQL, rightArgs := buildFunctionSQL(rightExpr.Operator, rightColumnName, rightExpr.Value)
-			if rightSQL == "" {
-				return fmt.Sprintf("CONCAT(%s, ?)", columnName), []interface{}{value}
-			}
-			return fmt.Sprintf("CONCAT(%s, %s)", columnName, rightSQL), rightArgs
-		}
-		// Check if value is a property name (string identifier)
-		if strVal, ok := value.(string); ok {
-			// Try to determine if it's a column name or a literal
-			// If it looks like a column name (starts with uppercase or contains underscore), treat as column
-			// Otherwise treat as literal
-			if len(strVal) > 0 && (strVal[0] >= 'A' && strVal[0] <= 'Z' || strings.Contains(strVal, "_")) {
-				// Treat as column name - convert to snake_case
-				columnName2 := toSnakeCase(strVal)
-				return fmt.Sprintf("CONCAT(%s, %s)", columnName, columnName2), nil
-			}
-		}
-		// Default: treat as literal
-		return fmt.Sprintf("CONCAT(%s, ?)", columnName), []interface{}{value}
-	case OpHas:
-		// Bitwise AND for enum flags: (column & value) = value
-		return fmt.Sprintf("(%s & ?)", columnName), []interface{}{value}
-	case OpSubstring:
-		return buildSubstringSQL(columnName, value)
-	case OpAdd:
-		return fmt.Sprintf("(%s + ?)", columnName), []interface{}{value}
-	case OpSub:
-		return fmt.Sprintf("(%s - ?)", columnName), []interface{}{value}
-	case OpMul:
-		return fmt.Sprintf("(%s * ?)", columnName), []interface{}{value}
-	case OpDiv:
-		return fmt.Sprintf("(%s / ?)", columnName), []interface{}{value}
-	case OpMod:
-		return fmt.Sprintf("(%s %% ?)", columnName), []interface{}{value}
-	// Date functions
-	case OpYear:
-		return fmt.Sprintf("CAST(strftime('%%Y', %s) AS INTEGER)", columnName), nil
-	case OpMonth:
-		return fmt.Sprintf("CAST(strftime('%%m', %s) AS INTEGER)", columnName), nil
-	case OpDay:
-		return fmt.Sprintf("CAST(strftime('%%d', %s) AS INTEGER)", columnName), nil
-	case OpHour:
-		return fmt.Sprintf("CAST(strftime('%%H', %s) AS INTEGER)", columnName), nil
-	case OpMinute:
-		return fmt.Sprintf("CAST(strftime('%%M', %s) AS INTEGER)", columnName), nil
-	case OpSecond:
-		return fmt.Sprintf("CAST(strftime('%%S', %s) AS INTEGER)", columnName), nil
-	case OpDate:
-		return fmt.Sprintf("DATE(%s)", columnName), nil
-	case OpTime:
-		return fmt.Sprintf("TIME(%s)", columnName), nil
-	case OpNow:
-		// now() is a parameterless function that returns current datetime
-		// columnName is ignored for now()
-		return "datetime('now')", nil
-	// Math functions
-	case OpCeiling:
-		// SQLite doesn't have native CEIL, so we implement it
-		return fmt.Sprintf("CASE WHEN %s = CAST(%s AS INTEGER) THEN %s ELSE CAST(%s AS INTEGER) + (CASE WHEN %s > 0 THEN 1 ELSE 0 END) END",
-			columnName, columnName, columnName, columnName, columnName), nil
-	case OpFloor:
-		// SQLite doesn't have native FLOOR, so we implement it
-		return fmt.Sprintf("CASE WHEN %s = CAST(%s AS INTEGER) THEN %s ELSE CAST(%s AS INTEGER) - (CASE WHEN %s < 0 THEN 1 ELSE 0 END) END",
-			columnName, columnName, columnName, columnName, columnName), nil
-	case OpRound:
-		return fmt.Sprintf("ROUND(%s)", columnName), nil
-	// Type conversion functions
-	case OpCast:
-		// value should be the type name (e.g., "Edm.String", "Edm.Int32")
-		if typeName, ok := value.(string); ok {
-			sqlType := edmTypeToSQLType(typeName)
-			return fmt.Sprintf("CAST(%s AS %s)", columnName, sqlType), nil
-		}
-		return "", nil
-	case OpIsOf:
-		// isof returns a boolean indicating if the value can be cast to the specified type
-		// In SQLite, we use a CASE expression to check if CAST succeeds
-		// Returns 1 (true) if cast is valid, 0 (false) otherwise
-		if typeName, ok := value.(string); ok {
-			// Check if this is an entity type check (no "Edm." prefix)
-			// For entity type checks on the current instance ($it), always return true
-			if len(typeName) > 0 && (len(typeName) < 4 || typeName[:4] != "Edm.") {
-				// Entity type check - for the current instance, always return 1 (true)
-				// because all records in the entity set are of that entity type
-				if columnName == "$it" {
-					return "1", nil
-				}
-				// For property-level entity type checks, we can't really validate at SQL level
-				// so we return 0 (false) as entity types don't apply to properties
-				return "0", nil
-			}
-
-			// EDM primitive type check
-			sqlType := edmTypeToSQLType(typeName)
-			// Use a safe type check that doesn't fail on invalid casts
-			// We try to cast and compare with the original to see if it's valid
-			return fmt.Sprintf("CASE WHEN CAST(%s AS %s) IS NOT NULL THEN 1 ELSE 0 END",
-				columnName, sqlType), nil
-		}
-		return "", nil
-	// Geospatial functions
-	case OpGeoDistance:
-		// geo.distance returns the distance in meters between two points
-		// The value contains the WKT string like "SRID=4326;POINT(0 0)"
-		// For SQLite with SpatiaLite extension, we would use:
-		// ST_Distance(geography_column, ST_GeomFromText(?))
-		// For basic implementation without spatial extensions, we return a placeholder
-		// that applications can override by implementing custom SQL functions
-		if geoStr, ok := value.(string); ok {
-			// Parse the WKT to extract coordinates if needed
-			// For now, we'll generate a function call that can be extended
-			return fmt.Sprintf("ST_Distance(%s, ST_GeomFromText(?, 4326))", columnName), []interface{}{geoStr}
-		}
-		return "", nil
-	case OpGeoLength:
-		// geo.length returns the length of a linestring in meters
-		// For SQLite with SpatiaLite: ST_Length(geography_column)
-		return fmt.Sprintf("ST_Length(%s)", columnName), nil
-	case OpGeoIntersects:
-		// geo.intersects returns true if two geometries intersect
-		// For SQLite with SpatiaLite: ST_Intersects(geography_column, ST_GeomFromText(?))
-		if geoStr, ok := value.(string); ok {
-			return fmt.Sprintf("ST_Intersects(%s, ST_GeomFromText(?, 4326))", columnName), []interface{}{geoStr}
-		}
-		return "", nil
-	default:
-		return "", nil
-	}
-}
-
-// edmTypeToSQLType converts OData EDM types to SQLite types
-func edmTypeToSQLType(edmType string) string {
-	switch edmType {
-	case "Edm.String":
-		return "TEXT"
-	case "Edm.Int32", "Edm.Int16", "Edm.Byte", "Edm.SByte":
-		return "INTEGER"
-	case "Edm.Int64":
-		return "INTEGER"
-	case "Edm.Decimal", "Edm.Double", "Edm.Single":
-		return "REAL"
-	case "Edm.Boolean":
-		return "INTEGER" // SQLite uses 0/1 for boolean
-	case "Edm.DateTimeOffset", "Edm.Date", "Edm.TimeOfDay":
-		return "TEXT" // SQLite stores dates as text
-	case "Edm.Guid":
-		return "TEXT"
-	case "Edm.Binary":
-		return "BLOB"
-	default:
-		return "TEXT" // Default to TEXT for unknown types
-	}
-}
-
-// buildSubstringSQL builds SQL for substring function
-// OData substring is 0-indexed, SQLite SUBSTR is 1-indexed
-// So we need to add 1 to the start position
-func buildSubstringSQL(columnName string, value interface{}) (string, []interface{}) {
-	args, ok := value.([]interface{})
-	if !ok {
-		return "", nil
-	}
-	if len(args) == 1 {
-		// substring(string, start) - from start to end
-		// Convert 0-indexed OData to 1-indexed SQLite by adding 1
-		return fmt.Sprintf("SUBSTR(%s, ? + 1, LENGTH(%s))", columnName, columnName), []interface{}{args[0]}
-	}
-	if len(args) == 2 {
-		// substring(string, start, length)
-		// Convert 0-indexed OData to 1-indexed SQLite by adding 1 to start
-		return fmt.Sprintf("SUBSTR(%s, ? + 1, ?)", columnName), args
-	}
-	return "", nil
-}
-
-// buildComparisonSQL builds SQL for comparison operator
-func buildComparisonSQL(op FilterOperator, leftSQL string) string {
-	switch op {
-	case OpEqual:
-		return fmt.Sprintf("%s = ?", leftSQL)
-	case OpNotEqual:
-		return fmt.Sprintf("%s != ?", leftSQL)
-	case OpGreaterThan:
-		return fmt.Sprintf("%s > ?", leftSQL)
-	case OpGreaterThanOrEqual:
-		return fmt.Sprintf("%s >= ?", leftSQL)
-	case OpLessThan:
-		return fmt.Sprintf("%s < ?", leftSQL)
-	case OpLessThanOrEqual:
-		return fmt.Sprintf("%s <= ?", leftSQL)
-	default:
-		return ""
-	}
-}
-
-// applyExpand applies expand (preload) options to the GORM query
-func applyExpand(db *gorm.DB, expand []ExpandOption, entityMetadata *metadata.EntityMetadata) *gorm.DB {
-	for _, expandOpt := range expand {
-		// Find the navigation property
-		navProp := findNavigationProperty(expandOpt.NavigationProperty, entityMetadata)
-		if navProp == nil {
-			continue // Skip invalid navigation property
-		}
-
-		// Build preload with conditions if nested options exist
-		if expandOpt.Select != nil || expandOpt.Filter != nil || expandOpt.OrderBy != nil ||
-			expandOpt.Top != nil || expandOpt.Skip != nil {
-			// Preload with conditions
-			db = db.Preload(navProp.Name, func(db *gorm.DB) *gorm.DB {
-				// Apply nested filter
-				if expandOpt.Filter != nil {
-					// Note: This would require navigation target metadata
-					// For now, just add basic support
-					db = applyFilterForExpand(db, expandOpt.Filter)
-				}
-
-				// Apply nested orderby
-				if len(expandOpt.OrderBy) > 0 {
-					for _, item := range expandOpt.OrderBy {
-						direction := "ASC"
-						if item.Descending {
-							direction = "DESC"
-						}
-						// Convert property name to snake_case for database column name
-						columnName := toSnakeCase(item.Property)
-						db = db.Order(fmt.Sprintf("%s %s", columnName, direction))
-					}
-				}
-
-				// Apply nested pagination
-				if expandOpt.Skip != nil {
-					db = db.Offset(*expandOpt.Skip)
-				}
-				if expandOpt.Top != nil {
-					db = db.Limit(*expandOpt.Top)
-				}
-
-				return db
-			})
-		} else {
-			// Simple preload without conditions
-			db = db.Preload(navProp.Name)
-		}
-	}
-	return db
-}
-
-// applyFilterForExpand applies filter to expanded navigation property
-// This is a simplified version that doesn't have full metadata context
-func applyFilterForExpand(db *gorm.DB, filter *FilterExpression) *gorm.DB {
-	if filter == nil {
-		return db
-	}
-
-	// Handle logical operators
-	if filter.Logical != "" {
-		leftDB := applyFilterForExpand(db, filter.Left)
-		rightDB := applyFilterForExpand(db, filter.Right)
-
-		switch filter.Logical {
-		case LogicalAnd:
-			return leftDB.Where(rightDB)
-		case LogicalOr:
-			leftQuery, leftArgs := buildSimpleFilterCondition(filter.Left)
-			rightQuery, rightArgs := buildSimpleFilterCondition(filter.Right)
-			combinedQuery := fmt.Sprintf("(%s) OR (%s)", leftQuery, rightQuery)
-			combinedArgs := append(leftArgs, rightArgs...)
-			return db.Where(combinedQuery, combinedArgs...)
-		}
-	}
-
-	// Handle simple comparison
-	query, args := buildSimpleFilterCondition(filter)
-	return db.Where(query, args...)
-}
-
-// buildSimpleFilterCondition builds a filter condition without metadata
-func buildSimpleFilterCondition(filter *FilterExpression) (string, []interface{}) {
-	if filter == nil {
-		return "", nil
-	}
-
-	if filter.Logical != "" {
-		leftQuery, leftArgs := buildSimpleFilterCondition(filter.Left)
-		rightQuery, rightArgs := buildSimpleFilterCondition(filter.Right)
-
-		switch filter.Logical {
-		case LogicalAnd:
-			query := fmt.Sprintf("(%s) AND (%s)", leftQuery, rightQuery)
-			args := append(leftArgs, rightArgs...)
-			return query, args
-		case LogicalOr:
-			query := fmt.Sprintf("(%s) OR (%s)", leftQuery, rightQuery)
-			args := append(leftArgs, rightArgs...)
-			return query, args
-		}
-	}
-
-	// Check if this is a function comparison (function on left side of comparison)
-	if filter.Left != nil && filter.Left.Operator != "" {
-		return buildSimpleFunctionComparison(filter)
-	}
-
-	// Convert property name to snake_case for database column name
-	fieldName := toSnakeCase(filter.Property)
-	return buildSimpleOperatorCondition(filter.Operator, fieldName, filter.Value)
-}
-
-// buildSimpleOperatorCondition builds SQL for a simple operator condition
-func buildSimpleOperatorCondition(op FilterOperator, fieldName string, value interface{}) (string, []interface{}) {
-	switch op {
-	case OpEqual:
-		return fmt.Sprintf("%s = ?", fieldName), []interface{}{value}
-	case OpNotEqual:
-		return fmt.Sprintf("%s != ?", fieldName), []interface{}{value}
-	case OpGreaterThan:
-		return fmt.Sprintf("%s > ?", fieldName), []interface{}{value}
-	case OpGreaterThanOrEqual:
-		return fmt.Sprintf("%s >= ?", fieldName), []interface{}{value}
-	case OpLessThan:
-		return fmt.Sprintf("%s < ?", fieldName), []interface{}{value}
-	case OpLessThanOrEqual:
-		return fmt.Sprintf("%s <= ?", fieldName), []interface{}{value}
-	case OpContains:
-		return fmt.Sprintf("%s LIKE ?", fieldName), []interface{}{"%" + fmt.Sprint(value) + "%"}
-	case OpStartsWith:
-		return fmt.Sprintf("%s LIKE ?", fieldName), []interface{}{fmt.Sprint(value) + "%"}
-	case OpEndsWith:
-		return fmt.Sprintf("%s LIKE ?", fieldName), []interface{}{"%" + fmt.Sprint(value)}
-	case OpHas:
-		// Bitwise AND for enum flags: (column & value) = value
-		return fmt.Sprintf("(%s & ?) = ?", fieldName), []interface{}{value, value}
-	default:
-		return "", nil
-	}
-}
-
-// buildSimpleFunctionComparison builds a comparison with a function on the left side (without metadata)
-func buildSimpleFunctionComparison(filter *FilterExpression) (string, []interface{}) {
-	funcExpr := filter.Left
-	fieldName := toSnakeCase(funcExpr.Property)
-
-	funcSQL, funcArgs := buildFunctionSQL(funcExpr.Operator, fieldName, funcExpr.Value)
-	if funcSQL == "" {
-		return "", nil
-	}
-
-	compSQL := buildComparisonSQL(filter.Operator, funcSQL)
-	if compSQL == "" {
-		return "", nil
-	}
-
-	allArgs := append(funcArgs, filter.Value)
-	return compSQL, allArgs
-}
-
-// applyOrderBy applies order by clauses to the GORM query
-func applyOrderBy(db *gorm.DB, orderBy []OrderByItem, entityMetadata *metadata.EntityMetadata) *gorm.DB {
-	for _, item := range orderBy {
-		// Check if this is a regular property (in entity metadata)
-		if propertyExists(item.Property, entityMetadata) {
-			// Regular entity property - use column name
-			columnName := GetColumnName(item.Property, entityMetadata)
-			// Use GORM clause.OrderByColumn to safely construct ORDER BY
-			db = db.Order(clause.OrderByColumn{
-				Column: clause.Column{Name: columnName},
-				Desc:   item.Descending,
-			})
-		} else {
-			// Not in entity metadata - assume it's a computed property alias
-			// Sanitize the alias to prevent SQL injection
-			sanitizedAlias := sanitizeIdentifier(item.Property)
-			if sanitizedAlias == "" {
-				continue // Skip invalid identifiers
-			}
-			db = db.Order(clause.OrderByColumn{
-				Column: clause.Column{Name: sanitizedAlias},
-				Desc:   item.Descending,
-			})
-		}
-	}
-	return db
-}
-
-// ApplySelect converts struct results to map format with only selected properties
-// This is called after the query to convert the result to the correct format for OData responses
-// The database query already fetched only the needed columns via applySelect
-func ApplySelect(results interface{}, selectedProperties []string, entityMetadata *metadata.EntityMetadata, expandOptions []ExpandOption) interface{} {
-	if len(selectedProperties) == 0 {
-		return results
-	}
-
-	// Get the slice value
-	sliceValue := reflect.ValueOf(results)
-	if sliceValue.Kind() != reflect.Slice {
-		return results
-	}
-
-	// Create a new slice of maps to hold the results
-	filteredResults := make([]map[string]interface{}, sliceValue.Len())
-
-	// Build a map of selected properties for quick lookup
-	selectedPropMap := make(map[string]bool)
-	// Also track navigation property selects (e.g., "Product/Name")
-	navPropSelects := make(map[string][]string) // nav prop name -> list of sub-properties
-
-	for _, propName := range selectedProperties {
-		propName = strings.TrimSpace(propName)
-		// Check if this is a navigation property path (e.g., "Product/Name")
-		if strings.Contains(propName, "/") {
-			parts := strings.SplitN(propName, "/", 2)
-			navProp := strings.TrimSpace(parts[0])
-			subProp := strings.TrimSpace(parts[1])
-			if navPropSelects[navProp] == nil {
-				navPropSelects[navProp] = []string{}
-			}
-			navPropSelects[navProp] = append(navPropSelects[navProp], subProp)
-		} else {
-			selectedPropMap[propName] = true
-		}
-	}
-
-	// Build a map of expanded properties for quick lookup
-	expandedPropMap := make(map[string]*ExpandOption)
-	for i := range expandOptions {
-		expandedPropMap[expandOptions[i].NavigationProperty] = &expandOptions[i]
-	}
-
-	// Always include key properties (OData requirement)
-	keyPropMap := make(map[string]bool)
-	for _, keyProp := range entityMetadata.KeyProperties {
-		keyPropMap[keyProp.Name] = true
-	}
-
-	for i := 0; i < sliceValue.Len(); i++ {
-		item := sliceValue.Index(i)
-		filteredItem := make(map[string]interface{})
-
-		// Extract selected properties and key properties
-		for _, prop := range entityMetadata.Properties {
-			// Skip complex types - they cannot be selected independently
-			if prop.IsComplexType {
-				continue
-			}
-
-			// Include if selected OR if it's a key property OR if it's an expanded navigation property
-			isSelected := selectedPropMap[prop.JsonName] || selectedPropMap[prop.Name]
-			isKey := keyPropMap[prop.Name]
-			isExpanded := prop.IsNavigationProp && (expandedPropMap[prop.Name] != nil || expandedPropMap[prop.JsonName] != nil)
-			hasNavSelect := len(navPropSelects[prop.JsonName]) > 0 || len(navPropSelects[prop.Name]) > 0
-
-			if isSelected || isKey || isExpanded || hasNavSelect {
-				// Get the field value
-				fieldValue := item.FieldByName(prop.Name)
-				if fieldValue.IsValid() && fieldValue.CanInterface() {
-					fieldVal := fieldValue.Interface()
-
-					// If this is an expanded navigation property with a nested select, apply it
-					if prop.IsNavigationProp && (isExpanded || hasNavSelect) {
-						var expandOpt *ExpandOption
-						if expandedPropMap[prop.Name] != nil {
-							expandOpt = expandedPropMap[prop.Name]
-						} else if expandedPropMap[prop.JsonName] != nil {
-							expandOpt = expandedPropMap[prop.JsonName]
-						}
-
-						// Get nested select from expand option or from navigation path selects
-						var nestedSelect []string
-						if expandOpt != nil && len(expandOpt.Select) > 0 {
-							nestedSelect = expandOpt.Select
-						} else if len(navPropSelects[prop.JsonName]) > 0 {
-							nestedSelect = navPropSelects[prop.JsonName]
-						} else if len(navPropSelects[prop.Name]) > 0 {
-							nestedSelect = navPropSelects[prop.Name]
-						}
-
-						// Apply nested select if specified
-						if len(nestedSelect) > 0 && fieldVal != nil {
-							fieldVal = applySelectToExpandedEntity(fieldVal, nestedSelect)
-						}
-					}
-
-					filteredItem[prop.JsonName] = fieldVal
-				}
-			}
-		}
-
-		filteredResults[i] = filteredItem
-	}
-
-	return filteredResults
-}
-
-// ApplySelectToEntity applies the $select filter to a single entity
-func ApplySelectToEntity(entity interface{}, selectedProperties []string, entityMetadata *metadata.EntityMetadata, expandOptions []ExpandOption) interface{} {
-	if len(selectedProperties) == 0 {
-		return entity
-	}
-
-	// Build a map of selected properties for quick lookup
-	selectedPropMap := make(map[string]bool)
-	// Also track navigation property selects (e.g., "Product/Name")
-	navPropSelects := make(map[string][]string) // nav prop name -> list of sub-properties
-
-	for _, propName := range selectedProperties {
-		propName = strings.TrimSpace(propName)
-		// Check if this is a navigation property path (e.g., "Product/Name")
-		if strings.Contains(propName, "/") {
-			parts := strings.SplitN(propName, "/", 2)
-			navProp := strings.TrimSpace(parts[0])
-			subProp := strings.TrimSpace(parts[1])
-			if navPropSelects[navProp] == nil {
-				navPropSelects[navProp] = []string{}
-			}
-			navPropSelects[navProp] = append(navPropSelects[navProp], subProp)
-		} else {
-			selectedPropMap[propName] = true
-		}
-	}
-
-	// Build a map of expanded properties for quick lookup
-	expandedPropMap := make(map[string]*ExpandOption)
-	for i := range expandOptions {
-		expandedPropMap[expandOptions[i].NavigationProperty] = &expandOptions[i]
-	}
-
-	// Get the entity value
-	entityValue := reflect.ValueOf(entity)
-	if entityValue.Kind() == reflect.Ptr {
-		entityValue = entityValue.Elem()
-	}
-
-	// Create a map to hold the filtered result
-	filteredEntity := make(map[string]interface{})
-
-	// Always include key properties (OData requirement)
-	keyPropMap := make(map[string]bool)
-	for _, keyProp := range entityMetadata.KeyProperties {
-		keyPropMap[keyProp.Name] = true
-	}
-
-	// Extract selected properties and key properties
-	for _, prop := range entityMetadata.Properties {
-		// Skip complex types - they cannot be selected independently
-		if prop.IsComplexType {
-			continue
-		}
-
-		// Include if selected OR if it's a key property OR if it's an expanded navigation property
-		isSelected := selectedPropMap[prop.JsonName] || selectedPropMap[prop.Name]
-		isKey := keyPropMap[prop.Name]
-		isExpanded := prop.IsNavigationProp && (expandedPropMap[prop.Name] != nil || expandedPropMap[prop.JsonName] != nil)
-		hasNavSelect := len(navPropSelects[prop.JsonName]) > 0 || len(navPropSelects[prop.Name]) > 0
-
-		if isSelected || isKey || isExpanded || hasNavSelect {
-			// Get the field value
-			fieldValue := entityValue.FieldByName(prop.Name)
-			if fieldValue.IsValid() && fieldValue.CanInterface() {
-				fieldVal := fieldValue.Interface()
-
-				// If this is an expanded navigation property with a nested select, apply it
-				if prop.IsNavigationProp && (isExpanded || hasNavSelect) {
-					var expandOpt *ExpandOption
-					if expandedPropMap[prop.Name] != nil {
-						expandOpt = expandedPropMap[prop.Name]
-					} else if expandedPropMap[prop.JsonName] != nil {
-						expandOpt = expandedPropMap[prop.JsonName]
-					}
-
-					// Get nested select from expand option or from navigation path selects
-					var nestedSelect []string
-					if expandOpt != nil && len(expandOpt.Select) > 0 {
-						nestedSelect = expandOpt.Select
-					} else if len(navPropSelects[prop.JsonName]) > 0 {
-						nestedSelect = navPropSelects[prop.JsonName]
-					} else if len(navPropSelects[prop.Name]) > 0 {
-						nestedSelect = navPropSelects[prop.Name]
-					}
-
-					// Apply nested select if specified
-					if len(nestedSelect) > 0 && fieldVal != nil {
-						fieldVal = applySelectToExpandedEntity(fieldVal, nestedSelect)
-					}
-				}
-
-				filteredEntity[prop.JsonName] = fieldVal
-			}
-		}
-	}
-
-	return filteredEntity
-}
-
-// applyTransformations applies apply transformations to the GORM query
-func applyTransformations(db *gorm.DB, transformations []ApplyTransformation, entityMetadata *metadata.EntityMetadata) *gorm.DB {
-	// Set the table/model for the query - use the statement destination if available
-	// Otherwise, we'll need the caller to set this
-	if db.Statement != nil && db.Statement.Dest == nil {
-		// Create an instance of the entity type to set the model
-		modelInstance := reflect.New(entityMetadata.EntityType).Interface()
-		db = db.Model(modelInstance)
-	}
-
-	// Track whether we've encountered a groupby/aggregate transformation
-	// to determine whether to use WHERE or HAVING for subsequent filters
-	hasGrouping := false
-
-	for _, transformation := range transformations {
-		switch transformation.Type {
-		case ApplyTypeGroupBy:
-			db = applyGroupBy(db, transformation.GroupBy, entityMetadata)
-			hasGrouping = true
-		case ApplyTypeAggregate:
-			db = applyAggregate(db, transformation.Aggregate, entityMetadata)
-			hasGrouping = true
-		case ApplyTypeFilter:
-			if hasGrouping {
-				// After groupby/aggregate, use HAVING clause
-				db = applyHavingFilter(db, transformation.Filter, entityMetadata)
-			} else {
-				// Before groupby/aggregate, use WHERE clause
-				db = applyFilter(db, transformation.Filter, entityMetadata)
-			}
-		case ApplyTypeCompute:
-			db = applyCompute(db, transformation.Compute, entityMetadata)
-		}
-	}
-	return db
-}
-
-// applyGroupBy applies a groupby transformation to the GORM query
-func applyGroupBy(db *gorm.DB, groupBy *GroupByTransformation, entityMetadata *metadata.EntityMetadata) *gorm.DB {
-	if groupBy == nil || len(groupBy.Properties) == 0 {
-		return db
-	}
-
-	// Build the GROUP BY clause
-	groupByColumns := make([]string, 0, len(groupBy.Properties))
-	selectColumns := make([]string, 0, len(groupBy.Properties))
-
-	for _, propName := range groupBy.Properties {
-		// Find the property metadata
-		prop := findProperty(propName, entityMetadata)
-		if prop == nil {
-			continue
-		}
-
-		// Get the database column name (snake_case)
-		columnName := GetColumnName(propName, entityMetadata)
-		groupByColumns = append(groupByColumns, columnName)
-		// Use "column as fieldname" to ensure the result has the correct JSON key
-		selectColumns = append(selectColumns, fmt.Sprintf("%s as %s", columnName, prop.JsonName))
-	}
-
-	// Apply nested transformations (typically aggregate)
-	if len(groupBy.Transform) > 0 {
-		for _, trans := range groupBy.Transform {
-			if trans.Type == ApplyTypeAggregate && trans.Aggregate != nil {
-				// Build aggregate expressions for SELECT
-				for _, aggExpr := range trans.Aggregate.Expressions {
-					aggSQL := buildAggregateSQL(aggExpr, entityMetadata)
-					if aggSQL != "" {
-						selectColumns = append(selectColumns, aggSQL)
-					}
-				}
-			}
-		}
-	} else {
-		// No nested transformations - add $count virtual property
-		selectColumns = append(selectColumns, "COUNT(*) as `$count`")
-	}
-
-	// Apply GROUP BY and SELECT
-	if len(groupByColumns) > 0 {
-		db = db.Group(strings.Join(groupByColumns, ", "))
-	}
-
-	if len(selectColumns) > 0 {
-		db = db.Select(strings.Join(selectColumns, ", "))
-	}
-
-	return db
-}
-
-// applyAggregate applies an aggregate transformation to the GORM query
-func applyAggregate(db *gorm.DB, aggregate *AggregateTransformation, entityMetadata *metadata.EntityMetadata) *gorm.DB {
-	if aggregate == nil || len(aggregate.Expressions) == 0 {
-		return db
-	}
-
-	// Build aggregate expressions for SELECT
-	selectColumns := make([]string, 0, len(aggregate.Expressions))
-	for _, aggExpr := range aggregate.Expressions {
-		aggSQL := buildAggregateSQL(aggExpr, entityMetadata)
-		if aggSQL != "" {
-			selectColumns = append(selectColumns, aggSQL)
-		}
-	}
-
-	if len(selectColumns) > 0 {
-		db = db.Select(strings.Join(selectColumns, ", "))
-	}
-
-	return db
-}
-
-// buildAggregateSQL builds the SQL for an aggregate expression
-func buildAggregateSQL(aggExpr AggregateExpression, entityMetadata *metadata.EntityMetadata) string {
-	// Handle $count special case
-	if aggExpr.Property == "$count" {
-		return fmt.Sprintf("COUNT(*) as `%s`", aggExpr.Alias)
-	}
-
-	// Find the property metadata
-	prop := findProperty(aggExpr.Property, entityMetadata)
-	if prop == nil {
-		return ""
-	}
-
-	// Build the aggregate SQL based on method
-	// Get the database column name (snake_case)
-	columnName := GetColumnName(aggExpr.Property, entityMetadata)
-
-	var sqlFunc string
-	switch aggExpr.Method {
-	case AggregationSum:
-		sqlFunc = "SUM"
-	case AggregationAvg:
-		sqlFunc = "AVG"
-	case AggregationMin:
-		sqlFunc = "MIN"
-	case AggregationMax:
-		sqlFunc = "MAX"
-	case AggregationCount:
-		sqlFunc = "COUNT"
-	case AggregationCountDistinct:
-		return fmt.Sprintf("COUNT(DISTINCT %s) as `%s`", columnName, aggExpr.Alias)
-	default:
-		return ""
-	}
-
-	return fmt.Sprintf("%s(%s) as `%s`", sqlFunc, columnName, aggExpr.Alias)
-}
-
-// applyCompute applies a compute transformation to the GORM query
-func applyCompute(db *gorm.DB, compute *ComputeTransformation, entityMetadata *metadata.EntityMetadata) *gorm.DB {
-	if compute == nil || len(compute.Expressions) == 0 {
-		return db
-	}
-
-	// Set the table/model for the query - similar to applyTransformations
-	if db.Statement != nil && db.Statement.Dest == nil {
-		// Create an instance of the entity type to set the model
-		modelInstance := reflect.New(entityMetadata.EntityType).Interface()
-		db = db.Model(modelInstance)
-	}
-
-	// Build SELECT clause with computed expressions
-	// We need to select all original columns plus computed columns
-	selectColumns := make([]string, 0)
-
-	// Build a set of field names to skip (stream auxiliary fields)
-	streamAuxFields := make(map[string]bool)
-	for _, streamProp := range entityMetadata.StreamProperties {
-		if streamProp.StreamContentTypeField != "" {
-			streamAuxFields[streamProp.StreamContentTypeField] = true
-		}
-		if streamProp.StreamContentField != "" {
-			streamAuxFields[streamProp.StreamContentField] = true
-		}
-	}
-
-	// Add all original entity properties
-	for _, prop := range entityMetadata.Properties {
-		// Skip navigation properties, complex types, stream properties, and stream auxiliary fields
-		if !prop.IsNavigationProp && !prop.IsComplexType && !prop.IsStream && !streamAuxFields[prop.FieldName] {
-			// Use snake_case column name for SQL, but alias as JsonName for the result
-			columnName := toSnakeCase(prop.Name)
-			selectColumns = append(selectColumns, fmt.Sprintf("%s as %s", columnName, prop.JsonName))
-		}
-	}
-
-	// Add computed expressions
-	for _, computeExpr := range compute.Expressions {
-		computeSQL := buildComputeSQL(computeExpr, entityMetadata)
-		if computeSQL != "" {
-			selectColumns = append(selectColumns, computeSQL)
-		}
-	}
-
-	if len(selectColumns) > 0 {
-		db = db.Select(strings.Join(selectColumns, ", "))
-	}
-
-	return db
-}
-
-// buildComputeSQL builds the SQL for a compute expression
-func buildComputeSQL(computeExpr ComputeExpression, entityMetadata *metadata.EntityMetadata) string {
-	if computeExpr.Expression == nil {
-		return ""
-	}
-
-	expr := computeExpr.Expression
-
-	// Check if this is a simple function call (no left/right expressions)
-	if expr.Left == nil && expr.Right == nil && expr.Operator != "" && expr.Property != "" {
-		// This is a simple function like year(CreatedAt)
-		prop := findProperty(expr.Property, entityMetadata)
-		if prop == nil {
-			return ""
-		}
-
-		// Use snake_case column name for SQL
-		columnName := toSnakeCase(prop.Name)
-		funcSQL, _ := buildFunctionSQL(expr.Operator, columnName, nil)
-		if funcSQL == "" {
-			return ""
-		}
-
-		return fmt.Sprintf("%s as %s", funcSQL, computeExpr.Alias)
-	}
-
-	// Handle arithmetic operations stored in Logical field (due to AST parser implementation)
-	// The AST parser stores arithmetic operators like "add", "sub", "mul", "div" in the Logical field
-	if expr.Left != nil && expr.Right != nil && expr.Logical != "" {
-		// Build SQL for left side (typically a property)
-		leftSQL := buildComputeExpressionSQL(expr.Left, entityMetadata)
-		if leftSQL == "" {
-			return ""
-		}
-
-		// Build SQL for right side (typically a literal value or another expression)
-		rightSQL := buildComputeExpressionSQL(expr.Right, entityMetadata)
-		if rightSQL == "" {
-			return ""
-		}
-
-		// Get the SQL operator from the Logical field
-		sqlOp := ""
-		switch expr.Logical {
-		case "add":
-			sqlOp = "+"
-		case "sub":
-			sqlOp = "-"
-		case "mul":
-			sqlOp = "*"
-		case "div":
-			sqlOp = "/"
-		case "mod":
-			sqlOp = "%"
-		default:
-			// Not an arithmetic operator (might be "and"/"or"), not supported here
-			return ""
-		}
-
-		return fmt.Sprintf("(%s %s %s) as %s", leftSQL, sqlOp, rightSQL, computeExpr.Alias)
-	}
-
-	// Handle arithmetic and other binary operations with operator set (legacy/alternative path)
-	if expr.Left != nil && expr.Right != nil && expr.Operator != "" {
-		// Build SQL for left side (typically a property)
-		leftSQL := buildComputeExpressionSQL(expr.Left, entityMetadata)
-		if leftSQL == "" {
-			return ""
-		}
-
-		// Build SQL for right side (typically a literal value or another expression)
-		rightSQL := buildComputeExpressionSQL(expr.Right, entityMetadata)
-		if rightSQL == "" {
-			return ""
-		}
-
-		// Get the SQL operator
-		sqlOp := ""
-		switch expr.Operator {
-		case OpAdd:
-			sqlOp = "+"
-		case OpSub:
-			sqlOp = "-"
-		case OpMul:
-			sqlOp = "*"
-		case OpDiv:
-			sqlOp = "/"
-		case OpMod:
-			sqlOp = "%"
-		default:
-			return ""
-		}
-
-		return fmt.Sprintf("(%s %s %s) as %s", leftSQL, sqlOp, rightSQL, computeExpr.Alias)
-	}
-
-	// For other complex expressions, not yet supported
-	return ""
-}
-
-// buildComputeExpressionSQL builds SQL for a sub-expression in a compute
-func buildComputeExpressionSQL(expr *FilterExpression, entityMetadata *metadata.EntityMetadata) string {
-	if expr == nil {
-		return ""
-	}
-
-	// Handle simple property reference (might have eq/true from AST parser)
-	if expr.Property != "" && expr.Left == nil && expr.Right == nil {
-		// This is a property reference, possibly with Operator=eq and Value=true from AST conversion
-		prop := findProperty(expr.Property, entityMetadata)
-		if prop == nil {
-			return ""
-		}
-		// Use snake_case column name for SQL
-		return toSnakeCase(prop.Name)
-	}
-
-	// Handle literal value
-	if expr.Value != nil && expr.Property == "" && expr.Left == nil && expr.Right == nil {
-		// Format the value appropriately for SQL
-		switch v := expr.Value.(type) {
-		case bool:
-			// For boolean literals in arithmetic, convert true to 1, false to 0
-			// But actually, this is a value like 1.1, so just return it
-			if v {
-				return "1"
-			}
-			return "0"
-		case string:
-			// String literals need quotes
-			return fmt.Sprintf("'%s'", v)
-		default:
-			// Numeric and other literals
-			return fmt.Sprintf("%v", expr.Value)
-		}
-	}
-
-	// Handle function call (e.g., toupper(Name))
-	// Must check Left/Right are nil AND Value is not just a boolean flag
-	if expr.Property != "" && expr.Operator != "" && expr.Operator != OpEqual && expr.Left == nil && expr.Right == nil {
-		prop := findProperty(expr.Property, entityMetadata)
-		if prop == nil {
-			return ""
-		}
-		// Use snake_case column name for SQL
-		columnName := toSnakeCase(prop.Name)
-		funcSQL, _ := buildFunctionSQL(expr.Operator, columnName, expr.Value)
-		return funcSQL
-	}
-
-	// Handle nested arithmetic (recursive case)
-	if expr.Left != nil && expr.Right != nil && expr.Operator != "" {
-		leftSQL := buildComputeExpressionSQL(expr.Left, entityMetadata)
-		rightSQL := buildComputeExpressionSQL(expr.Right, entityMetadata)
-		if leftSQL == "" || rightSQL == "" {
-			return ""
-		}
-
-		sqlOp := ""
-		switch expr.Operator {
-		case OpAdd:
-			sqlOp = "+"
-		case OpSub:
-			sqlOp = "-"
-		case OpMul:
-			sqlOp = "*"
-		case OpDiv:
-			sqlOp = "/"
-		case OpMod:
-			sqlOp = "%"
-		default:
-			return ""
-		}
-
-		return fmt.Sprintf("(%s %s %s)", leftSQL, sqlOp, rightSQL)
-	}
-
-	// Handle nested arithmetic in Logical field (alternative encoding)
-	if expr.Left != nil && expr.Right != nil && expr.Logical != "" {
-		leftSQL := buildComputeExpressionSQL(expr.Left, entityMetadata)
-		rightSQL := buildComputeExpressionSQL(expr.Right, entityMetadata)
-		if leftSQL == "" || rightSQL == "" {
-			return ""
-		}
-
-		sqlOp := ""
-		switch expr.Logical {
-		case "add":
-			sqlOp = "+"
-		case "sub":
-			sqlOp = "-"
-		case "mul":
-			sqlOp = "*"
-		case "div":
-			sqlOp = "/"
-		case "mod":
-			sqlOp = "%"
-		default:
-			return ""
-		}
-
-		return fmt.Sprintf("(%s %s %s)", leftSQL, sqlOp, rightSQL)
-	}
-
-	return ""
-}
-
-// findProperty finds a property by name or JSON name in the entity metadata
-func findProperty(propName string, entityMetadata *metadata.EntityMetadata) *metadata.PropertyMetadata {
-	if entityMetadata == nil {
-		return nil
-	}
-	return entityMetadata.FindProperty(propName)
-}
-
-// sanitizeIdentifier sanitizes a user-provided identifier to prevent SQL injection.
-// It ensures that the identifier contains only safe characters (alphanumeric and underscore).
-// Returns an empty string if the identifier contains invalid characters.
-// Special OData identifiers like "$it" are allowed.
-func sanitizeIdentifier(identifier string) string {
-	if identifier == "" {
-		return ""
-	}
-
-	// Allow special OData reserved identifiers
-	if identifier == "$it" || identifier == "$count" {
-		return identifier
-	}
-
-	// Allow OData identifiers starting with $
-	if len(identifier) > 1 && identifier[0] == '$' {
-		// Ensure the rest is alphanumeric/underscore
-		for i, ch := range identifier[1:] {
-			if i == 0 {
-				// Second character must be a letter or underscore
-				if (ch < 'a' || ch > 'z') && (ch < 'A' || ch > 'Z') && ch != '_' {
-					return ""
-				}
-			} else {
-				// Subsequent characters can be alphanumeric or underscore
-				if (ch < 'a' || ch > 'z') && (ch < 'A' || ch > 'Z') && (ch < '0' || ch > '9') && ch != '_' {
-					return ""
-				}
-			}
-		}
-		// Valid $ identifier
-		return identifier
-	}
-
-	// Check each character - only allow alphanumeric, underscore, and must start with letter or underscore
-	for i, ch := range identifier {
-		if i == 0 {
-			// First character must be a letter or underscore
-			if (ch < 'a' || ch > 'z') && (ch < 'A' || ch > 'Z') && ch != '_' {
-				return ""
-			}
-		} else {
-			// Subsequent characters can be alphanumeric or underscore
-			if (ch < 'a' || ch > 'z') && (ch < 'A' || ch > 'Z') && (ch < '0' || ch > '9') && ch != '_' {
-				return ""
-			}
-		}
-	}
-
-	// Check for SQL reserved keywords that could be dangerous
-	upper := strings.ToUpper(identifier)
-	reservedKeywords := []string{
-		"SELECT", "INSERT", "UPDATE", "DELETE", "DROP", "CREATE", "ALTER", "TRUNCATE",
-		"UNION", "JOIN", "WHERE", "FROM", "INTO", "VALUES", "SET", "EXEC", "EXECUTE",
-	}
-	for _, keyword := range reservedKeywords {
-		if upper == keyword {
-			return ""
-		}
-	}
-
-	return identifier
 }
